Follow Sub Rate Limit Workaround
I was running into issues with the rate limiting of the Twitch API while
putting follow requests, so built in a ratelimit and popped in some
setInterval timing to make it work. Still going to have compare current
follows with followers so that we don't try to follow people we're
already following.